fix bug in profile page - #2
